Add extra warning flags to compiler options
- Add -W to compiler options for library and test suite.  Disable
unused-parameter and missing-field-initializers checks, since these are
valid behaviors that we use a lot.

- Fix resulting warnings

+++ lib/connection.c
@@ -788,7 +788,7 @@ exported int mrpc_conn_set_operations(struct mrpc_connection *conn,
 exported int mrpc_conn_get_counter(struct mrpc_connection *conn,
 			enum mrpc_conn_counter counter, uint64_t *result)
 {
-	if (conn == NULL || result == NULL || counter < 0 ||
+	if (conn == NULL || result == NULL || (int)counter < 0 ||
 				counter >= MRPC_CONNCTR_NR)
 		return EINVAL;
 	pthread_mutex_lock(&conn->counters_lock);
